+++ data/src/main/java/com/github/lexleontiev/chatexample/data/ChatRepository.kt
@@ -1,0 +1,10 @@
+package com.github.lexleontiev.chatexample.data
+
+import kotlinx.coroutines.flow.Flow
+
+
+interface ChatRepository {
+    fun getMessages(): Flow<List<Message>>
+    suspend fun sendMessage(message: Message)
+    suspend fun removeAllMessages()
+}
