+++ data/src/main/java/com/github/lexleontiev/chatexample/data/room/AppDatabase.kt
@@ -1,0 +1,34 @@
+package com.github.lexleontiev.chatexample.data.room
+
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
+import android.content.Context
+
+
+@Database(
+    entities = [MessageDTO::class],
+    version = 1,
+    exportSchema = false
+)
+abstract class AppDatabase : RoomDatabase() {
+
+    abstract fun messageDao(): MessageDao
+
+    companion object {
+        @Volatile
+        private var INSTANCE: AppDatabase? = null
+
+        fun getDatabase(context: Context): AppDatabase {
+            return INSTANCE ?: synchronized(this) {
+                val instance = Room.databaseBuilder(
+                    context.applicationContext,
+                    AppDatabase::class.java,
+                    "chatexample_database"
+                ).build()
+                INSTANCE = instance
+                instance
+            }
+        }
+    }
+}
