+++ data/src/main/java/com/github/lexleontiev/chatexample/data/room/MessageDao.kt
@@ -1,0 +1,24 @@
+package com.github.lexleontiev.chatexample.data.room
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import kotlinx.coroutines.flow.Flow
+
+
+@Dao
+interface MessageDao {
+
+    @Query("SELECT * FROM messages ORDER BY timestamp ASC")
+    fun getAllMessages(): Flow<List<MessageDTO>>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertMessage(message: MessageDTO)
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertMessages(messages: List<MessageDTO>)
+
+    @Query("DELETE FROM messages")
+    suspend fun deleteAllMessages()
+}
