+++ feature/chat/build.gradle.kts
@@ -1,0 +1,72 @@
+plugins {
+    id("com.android.library")
+    alias(libs.plugins.compose.compiler)
+    kotlin("android")
+    alias(libs.plugins.ksp)
+    alias(libs.plugins.hilt)
+}
+
+android {
+    compileSdk = libs.versions.compile.sdk.version.get().toInt()
+
+    defaultConfig {
+        minSdk = libs.versions.min.sdk.version.get().toInt()
+        namespace = "com.github.lexleontiev.chatexample.feature.chat"
+
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+
+        vectorDrawables {
+            useSupportLibrary = true
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+
+    kotlinOptions {
+        jvmTarget = JavaVersion.VERSION_17.toString()
+    }
+
+    buildFeatures {
+        compose = true
+        viewBinding = true
+        buildConfig = false
+    }
+
+    composeOptions {
+        kotlinCompilerExtensionVersion = libs.versions.compose.compilerextension.get()
+    }
+
+    lint {
+        warningsAsErrors = true
+        abortOnError = true
+        disable.add("GradleDependency")
+    }
+}
+
+dependencies {
+    implementation(projects.data)
+    implementation(libs.androidx.appcompat)
+    implementation(libs.androidx.core.ktx)
+
+    implementation(platform(libs.compose.bom))
+    implementation(libs.androidx.activity.compose)
+    implementation(libs.compose.ui)
+    implementation(libs.compose.ui.tooling)
+    implementation(libs.compose.foundation)
+    implementation(libs.compose.material)
+    implementation(libs.lifecycle.runtime.compose)
+
+    implementation(libs.hilt.android)
+    implementation(libs.hilt.navigation.compose)
+    ksp(libs.hilt.compiler)
+
+    testImplementation(libs.junit)
+
+    debugImplementation(libs.compose.ui.test.manifest)
+    androidTestImplementation(libs.compose.ui.test.junit4)
+    androidTestImplementation(libs.androidx.test.runner)
+    androidTestImplementation(libs.androidx.test.ext.junit)
+}
