+++ feature/chat/src/main/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <application>
+        <activity
+            android:name="com.github.lexleontiev.chatexample.feature.chat.ChatActivity"
+            android:windowSoftInputMode="adjustResize"
+            android:exported="true" />
+    </application>
+
+</manifest>
