+++ feature/chat/src/main/java/com/github/lexleontiev/chatexample/feature/chat/ThemePreviews.kt
@@ -1,0 +1,9 @@
+package com.github.lexleontiev.chatexample.feature.chat
+
+import android.content.res.Configuration.UI_MODE_NIGHT_NO
+import android.content.res.Configuration.UI_MODE_NIGHT_YES
+import androidx.compose.ui.tooling.preview.Preview
+
+@Preview(name = "Dark Mode", showBackground = true, uiMode = UI_MODE_NIGHT_YES, backgroundColor = 0x00000000)
+@Preview(name = "Light Mode", showBackground = true, uiMode = UI_MODE_NIGHT_NO)
+annotation class ThemePreviews
