+++ feature/chat/src/main/java/com/github/lexleontiev/chatexample/feature/chat/components/MessageInput.kt
@@ -1,0 +1,117 @@
+package com.github.lexleontiev.chatexample.feature.chat.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.TextField
+import androidx.compose.material.TextFieldDefaults
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.Send
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.RectangleShape
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.unit.dp
+import com.github.lexleontiev.chatexample.feature.chat.ThemePreviews
+
+
+@Composable
+internal fun MessageInput(onSend: (String) -> Unit) {
+    var textState by remember { mutableStateOf(TextFieldValue("")) }
+    var isFocused by remember { mutableStateOf(false) }
+    val hasText by remember {
+        derivedStateOf {
+            textState.text.isNotEmpty()
+        }
+    }
+
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentHeight()
+            .shadow(8.dp, shape = RectangleShape)
+            .background(Color.White)
+            .padding(horizontal = 12.dp, vertical = 20.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        TextField(
+            value = textState,
+            onValueChange = { textState = it },
+            modifier = Modifier
+                .weight(1f)
+                .height(56.dp)
+                .border(
+                    width = 2.dp,
+                    color = if (isFocused) Color(0xFFFF4081) else Color(0xFFBDBDBD),
+                    shape = RoundedCornerShape(24.dp)
+                )
+                .padding(horizontal = 16.dp)
+                .clip(RoundedCornerShape(24.dp))
+                .onFocusChanged { focusState ->
+                    isFocused = focusState.isFocused
+                },
+            colors = TextFieldDefaults.textFieldColors(
+                backgroundColor = Color.White,
+                focusedIndicatorColor = Color.Transparent,
+                unfocusedIndicatorColor = Color.Transparent,
+                disabledIndicatorColor = Color.Transparent
+            )
+        )
+
+        Spacer(modifier = Modifier.width(16.dp))
+
+        Box(
+            modifier = Modifier
+                .size(56.dp)
+                .alpha(if (hasText) 1f else 0.5f)
+                .clip(CircleShape)
+                .background(
+                    brush = Brush.linearGradient(
+                        colors = listOf(Color(0xFFFF8A80), Color(0xFFFF4081))
+                    )
+                )
+                .clickable(hasText) {
+                    onSend(textState.text)
+                    textState = TextFieldValue("")
+                },
+            contentAlignment = Alignment.Center
+        ) {
+            Icon(
+                imageVector = Icons.AutoMirrored.Filled.Send,
+                contentDescription = "Send",
+                tint = Color.White
+            )
+        }
+    }
+}
+
+@ThemePreviews
+@Composable
+private fun MessageInputPreview() = MaterialTheme {
+    MessageInput {}
+}
