--- library-android/build.gradle.kts
@@ -1,71 +1,0 @@
-version = LibraryAndroidCoordinates.LIBRARY_VERSION
-
-plugins {
-    id("com.android.library")
-    kotlin("android")
-    id("maven-publish")
-    alias(libs.plugins.ksp)
-    alias(libs.plugins.hilt)
-    publish
-}
-
-android {
-    compileSdk = libs.versions.compile.sdk.version.get().toInt()
-
-    defaultConfig {
-        minSdk = libs.versions.min.sdk.version.get().toInt()
-        namespace = "com.github.lexleontiev.chatexample.library.android"
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-
-    kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_17.toString()
-    }
-
-    buildTypes {
-        getByName("release") {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-
-    publishing {
-        singleVariant("release") {
-            withSourcesJar()
-            withJavadocJar()
-        }
-    }
-
-    lint {
-        warningsAsErrors = true
-        abortOnError = true
-        disable.add("GradleDependency")
-    }
-}
-
-dependencies {
-    implementation(libs.androidx.appcompat)
-    implementation(libs.androidx.core.ktx)
-
-    implementation(libs.room.runtime)
-    implementation(libs.room.ktx)
-    ksp(libs.room.compiler)
-
-    implementation(libs.hilt.android)
-    ksp(libs.hilt.compiler)
-
-    testImplementation(libs.junit)
-
-    androidTestImplementation(libs.androidx.test.runner)
-    androidTestImplementation(libs.androidx.test.ext.junit)
-}
