--- library-android/src/main/java/com/github/lexleontiev/chatexample/library/android/ChatRepository.kt
@@ -1,10 +1,0 @@
-package com.github.lexleontiev.chatexample.library.android
-
-import kotlinx.coroutines.flow.Flow
-
-
-interface ChatRepository {
-    fun getMessages(): Flow<List<Message>>
-    suspend fun sendMessage(message: Message)
-    suspend fun removeAllMessages()
-}
