--- library-android/src/main/java/com/github/lexleontiev/chatexample/library/android/MemoryCache.kt
@@ -1,23 +1,0 @@
-package com.github.lexleontiev.chatexample.library.android
-
-import javax.inject.Inject
-
-
-class MemoryCache @Inject constructor() {
-
-    private val messages: MutableList<Message> = mutableListOf()
-
-    fun getMessages(): List<Message> = messages
-
-    fun saveMessage(message: Message) {
-        messages.add(message)
-    }
-
-    fun saveMessages(newMessages: List<Message>) {
-        messages.addAll(newMessages)
-    }
-
-    fun removeAllMessages() {
-        messages.clear()
-    }
-}
