--- library-compose/src/main/AndroidManifest.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application>
-        <activity
-            android:name="com.github.lexleontiev.chatexample.app.ui.ComposeActivity"
-            android:windowSoftInputMode="adjustResize"
-            android:exported="true" />
-    </application>
-
-</manifest>
