--- library-compose/src/main/java/com/github/lexleontiev/chatexample/app/ui/ThemePreviews.kt
@@ -1,9 +1,0 @@
-package com.github.lexleontiev.chatexample.app.ui
-
-import android.content.res.Configuration.UI_MODE_NIGHT_NO
-import android.content.res.Configuration.UI_MODE_NIGHT_YES
-import androidx.compose.ui.tooling.preview.Preview
-
-@Preview(name = "Dark Mode", showBackground = true, uiMode = UI_MODE_NIGHT_YES, backgroundColor = 0x00000000)
-@Preview(name = "Light Mode", showBackground = true, uiMode = UI_MODE_NIGHT_NO)
-annotation class ThemePreviews
