--- library-compose/src/main/java/com/github/lexleontiev/chatexample/app/ui/chat/ChatAppBar.kt
@@ -1,114 +1,0 @@
-package com.github.lexleontiev.chatexample.app.ui.chat
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.DropdownMenu
-import androidx.compose.material.DropdownMenuItem
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.MoreVert
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.PopupProperties
-import androidx.hilt.navigation.compose.hiltViewModel
-import com.github.lexleontiev.chatexample.app.ui.ThemePreviews
-import com.github.lexleontiev.chatexample.library.compose.R
-
-
-@Composable
-internal fun ChatAppBar(
-    modifier: Modifier = Modifier,
-    viewModel: ChatViewModel = hiltViewModel()
-) {
-    var expanded by remember { mutableStateOf(false) }
-    TopAppBar(
-        modifier = modifier.fillMaxWidth(),
-        title = {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Image(
-                    painter = painterResource(id = R.drawable.profile),
-                    contentDescription = "User Avatar",
-                    modifier = Modifier
-                        .size(40.dp)
-                        .padding(end = 8.dp),
-                    contentScale = ContentScale.Fit
-                )
-                Text(
-                    text = "Sarah",
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 18.sp
-                )
-            }
-        },
-        navigationIcon = {
-            IconButton(onClick = { /* no-op */ }) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = "Back"
-                )
-            }
-        },
-        actions = {
-            IconButton(onClick = { expanded = true }) {
-                Icon(
-                    imageVector = Icons.Filled.MoreVert,
-                    contentDescription = "Menu",
-                    modifier = Modifier.rotate(90f)
-                )
-            }
-            DropdownMenu(
-                expanded = expanded,
-                onDismissRequest = { expanded = false },
-                properties = PopupProperties(
-                    dismissOnBackPress = true,
-                    dismissOnClickOutside = true
-                )
-            ) {
-                DropdownMenuItem(
-                    onClick = {
-                        expanded = false
-                        viewModel.debugReceivingMessage()
-                    }) {
-                    Text("Receive message")
-                }
-                DropdownMenuItem(
-                    onClick = {
-                        expanded = false
-                        viewModel.debugClearChat()
-                    }) {
-                    Text("Clear chat")
-                }
-            }
-        },
-        backgroundColor = Color.White,
-        contentColor = Color(0xFF374655),
-        elevation = 4.dp
-    )
-}
-
-
-@ThemePreviews
-@Composable
-private fun ChatAppBarPreview() = MaterialTheme {
-    ChatAppBar()
-}
